+++ lib/rubocop/cop/rspec/rails/have_http_status.rb
@@ -9,6 +9,7 @@ module RuboCop
         # @example
         #   # bad
         #   expect(response.status).to be(200)
+        #   expect(response.code).to eq("200")
         #
         #   # good
         #   expect(response).to have_http_status(200)
@@ -18,7 +19,7 @@ module RuboCop
 
           MSG =
             'Prefer `expect(response).%<to>s have_http_status(%<status>i)` ' \
-            'over `expect(response.status).%<to>s %<match>s`.'
+            'over `%<bad_code>s`.'
 
           RUNNERS = %i[to to_not not_to].to_set
           RESTRICT_ON_SEND = RUNNERS
@@ -27,19 +28,21 @@ module RuboCop
           def_node_matcher :match_status, <<-PATTERN
             (send
               (send nil? :expect
-                $(send (send nil? :response) :status)
+                $(send (send nil? :response) {:status :code})
               )
               $RUNNERS
-              $(send nil? {:be :eq :eql :equal} (int $_))
+              $(send nil? {:be :eq :eql :equal} ({int str} $_))
             )
           PATTERN
 
           def on_send(node)
             match_status(node) do |response_status, to, match, status|
-              message = format(MSG, to: to, match: match.source, status: status)
+              message = format(MSG, to: to, status: status,
+                                    bad_code: node.source)
               add_offense(node, message: message) do |corrector|
                 corrector.replace(response_status, 'response')
                 corrector.replace(match.loc.selector, 'have_http_status')
+                corrector.replace(match.first_argument, status.to_i.to_s)
               end
             end
           end